#12 calendar 클릭 이벤트 - 데이터 있는 날짜 클릭씨 글출력( json출력 확인)
