+++ components/jobDiary/MainDiary.vue
@@ -1,0 +1,55 @@
+<template>
+    <div id="demo">
+        {{ diaryInfo }}
+    </div>
+</template>
+
+<script>
+export default {
+    props: ['diaryInfo'],
+    date(){
+      return {
+        
+      }  
+    },
+    methods: {
+        setMenu: function(top, left) {
+            
+            largestHeight = window.innerHeight - this.$$.right.offsetHeight - 25;
+            largestWidth = window.innerWidth - this.$$.right.offsetWidth - 25;
+
+            if (top > largestHeight) top = largestHeight;
+
+            if (left > largestWidth) left = largestWidth;
+
+            this.top = top + 'px';
+            this.left = left + 'px';
+        },
+
+        closeMenu: function() {
+            this.viewMenu = false;
+        },
+
+        openMenu: function(e) {
+            this.viewMenu = true;
+
+            Vue.nextTick(function() {
+                this.$$.right.focus();
+
+                this.setMenu(e.y, e.x)
+            }.bind(this));
+            e.preventDefault();
+        }
+    }
+
+}
+      
+
+   
+    
+</script>
+
+<style>
+
+</style>
+
